feat(materiales): 154 materiales del juego con textura del paquete real
Añade el catálogo clásico de materiales de Minecraft cargando las texturas
PNG reales del paquete (textures/blocks/) en el atlas, en vez de pintarlas
proceduralmente a mano. Para dar cabida a tantos bloques:

- Celda del mundo a 16 bits (Uint16Array): rompe el techo de 256 ids.
  Toca worldgen, el receptor del worker, el RLE de world.js (pares de 16
  bits) y storage (DB v2, migración que descarta guardados v1).
- Atlas 16x16 -> 32x32 (512px, ~1024 téselas). buildAtlasAsync carga los
  PNG y aplica los painters procedurales encima; boot() ahora es async.
- Generador tools/gen-materiales.mjs (tabla curada + validación de que
  cada PNG existe) -> js/materiales.data.js. js/materiales.js asigna ids
  (base 128) y téselas (base 200) de forma determinista; blocks.js y
  atlas.pngtiles.js consumen ese plan.

Familias: piedras y derivados, 5 maderas completas + troncos pelados,
minerales y bloques metálicos, y tintados x16 (hormigón, polvo, terracota,
terracota glaseada, cristal). Verificado en WebGL real (Edge headless):
0 sin registrar, texturas con color auténtico, arranque sin errores.

Incluye también el trabajo en curso de shatter/carve (desgranado de
bloques en mini-vóxeles) que estaba sin commitear en el working tree:
js/carve.js, js/shatter.js, sus tests y los enganches en audio/mesher/
renderer/main.

+++ js/worldgen.js
@@ -117,10 +117,11 @@ export class Generator {
     /**
      * Genera el chunk (cx, cz) completo. Función pura: el resultado depende
      * solo de (semilla, cx, cz).
-     * @returns {Uint8Array} CHUNK×SY×CHUNK, índice (y·CHUNK + lz)·CHUNK + lx
+     * @returns {Uint16Array} CHUNK×SY×CHUNK, índice (y·CHUNK + lz)·CHUNK + lx
+     *   (16 bits por celda para admitir más de 256 ids de bloque)
      */
     generateChunk(cx, cz) {
-        const blocks = new Uint8Array(CHUNK * SY * CHUNK);
+        const blocks = new Uint16Array(CHUNK * SY * CHUNK);
         const li = (lx, y, lz) => (y * CHUNK + lz) * CHUNK + lx;
         const x0 = cx * CHUNK, z0 = cz * CHUNK;
